requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
click==8.1.8
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.5
python-jose==3.4.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
tomli==2.2.1
typing_extensions==4.12.2
uvicorn==0.34.0
